backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Session
itsdangerous==2.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.19.0

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0
